take out mysql tests
