+++ frameworks/Java/act-beetlsql/pom.xml
@@ -21,6 +21,7 @@
         <HikariCP.version>2.6.1</HikariCP.version>
         <act-mustache.version>1.0.0</act-mustache.version>
         <osgl-tool.version>1.0.3</osgl-tool.version>
+        <beetlsql.version>2.8.9</beetlsql.version>
         <app.entry>com.techempower.act.AppEntry</app.entry>
     </properties>
 
@@ -64,6 +65,11 @@
             <groupId>org.osgl</groupId>
             <artifactId>osgl-tool</artifactId>
             <version>${osgl-tool.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.ibeetl</groupId>
+            <artifactId>beetlsql</artifactId>
+            <version>${beetlsql.version}</version>
         </dependency>
     </dependencies>
