--- frameworks/Java/act-beetlsql/src/main/java/com/techempower/act/domain/IWorld.java
@@ -1,12 +1,0 @@
-package com.techempower.act.domain;
-
-import act.util.SimpleBean;
-
-/**
- * We need this interface because we want to support both MongoDB and MySQL in the same application
- */
-public interface IWorld extends SimpleBean {
-    Integer getId();
-    Integer getRandomNumber();
-    void setRandomNumber(Integer randomNumber);
-}
