--- frameworks/Java/act-beetlsql/src/main/java/com/techempower/act/pgsql/domain/Fortune.java
@@ -1,18 +1,0 @@
-package com.techempower.act.pgsql.domain;
-
-
-import act.db.DB;
-import org.beetl.sql.core.mapper.BaseMapper;
-
-import javax.persistence.Entity;
-
-@DB("pgsql")
-@Entity
-public class Fortune extends com.techempower.act.sql.domain.Fortune {
-
-    public Fortune(Integer id, String message) {
-        super(id, message);
-    }
-
-    public static interface Dao extends BaseMapper<Fortune> {}
-}
