--- frameworks/Java/act-beetlsql/src/main/java/com/techempower/act/pgsql/domain/World.java
@@ -1,18 +1,0 @@
-package com.techempower.act.pgsql.domain;
-
-
-import act.db.DB;
-
-import javax.persistence.Entity;
-
-@DB("pgsql")
-@Entity
-public class World extends com.techempower.act.sql.domain.World {
-
-    public World(Integer id, Integer randomNumber) {
-        super(id, randomNumber);
-    }
-
-    public interface Dao extends WorldDao<World> {
-    }
-}
