+++ frameworks/Java/act-beetlsql/src/main/resources/conf/common/app.properties
@@ -1,9 +1,6 @@
 # follow Java world convention
 http.port=8080
 
-# disable charset output for JSON response
-render.json.output_charset=false
-
 # disable CLI service
 cli=false
