+++ frameworks/Java/act-beetlsql/src/main/resources/conf/common/db.properties
@@ -1,27 +1,12 @@
-db.instances=mysql,pgsql
+host=localhost
 
-mysql.host=localhost
-
-db.mysql.impl=act.db.beetlsql.BeetlSqlPlugin
-db.mysql.url=jdbc:mysql://${mysql.host}:3306/hello_world?jdbcCompliantTruncation=false&elideSetAutoCommits=true&useLocalSessionState=true&cachePrepStmts=true&cacheCallableStmts=true&alwaysSendSetIsolation=false&prepStmtCacheSize=4096&cacheServerConfiguration=true&prepStmtCacheSqlLimit=2048&zeroDateTimeBehavior=convertToNull&traceProtocol=false&useUnbufferedInput=false&useReadAheadInput=false&maintainTimeStats=false&useServerPrepStmts=true&cacheRSMetadata=true&serverTimezone=UTC
-db.mysql.username=benchmarkdbuser
-db.mysql.password=benchmarkdbpass
-db.mysql.maxConnections=256
-db.mysql.minConnections=128
-db.mysql.useServerPrepStmts=true
-db.mysql.naming.convention=underscore
-db.mysql.interceptor.debug=true
-db.mysql.ddl.create=false
-
-pgsql.host=localhost
-
-db.pgsql.impl=act.db.beetlsql.BeetlSqlPlugin
-db.pgsql.url=jdbc:postgresql://${pgsql.host}:5432/hello_world
-db.pgsql.username=benchmarkdbuser
-db.pgsql.password=benchmarkdbpass
-db.pgsql.namingconvention=com.avaje.ebean.config.MatchingNamingConvention
-db.pgsql.maxConnections=256
-db.pgsql.minConnections=128
-db.pgsql.naming.convention=underscore
-db.pgsql.interceptor.debug=false
-db.pgsql.ddl.create=false
+db.impl=act.db.beetlsql.BeetlSqlPlugin
+db.url=jdbc:postgresql://${host}:5432/hello_world
+db.username=benchmarkdbuser
+db.password=benchmarkdbpass
+db.namingconvention=com.avaje.ebean.config.MatchingNamingConvention
+db.maxConnections=256
+db.minConnections=128
+db.beetlsql.nc=org.beetl.sql.core.DefaultNameConversion
+db.interceptor.debug=false
+db.ddl.create=false
