+++ frameworks/Java/act-beetlsql/src/main/resources/conf/local/db.properties
@@ -1,2 +1,1 @@
-db.mysql.maxConnections=40
-db.pgsql.maxConnections=40
+db.maxConnections=40
