--- frameworks/Java/act-beetlsql/src/main/resources/routes.conf
@@ -1,1 +1,0 @@
-GET /plaintext echo: Hello, World!
